🔨 refactor: style 객체 분리
